Split the plugin into turn_taking/ submodules
Break the monolithic __init__.py into service / state / delivery / core /
patching / hooks under turn_taking/, leaving __init__.py as registration
only. Patch installers renamed to _patch_<original>; shared runtime state
made public in state.py.

Also key the epoch<->turn map by the raw event.message_id (bound via
_reply_anchor_for_event) instead of the reply anchor, which is None on
Telegram forum-topic groups -- so reply correlation holds across platforms.
